Validation notebooks

+++ Training/global_model/awk_data.py
@@ -123,7 +123,7 @@ def get_tensors_spec(config):
         "wind_X_norm" : tf.TensorSpec(shape=(None,len(config.columns["window_features"])), dtype=tf.float32),  #windox_X (batch, #wind_x)
         "wind_meta": tf.TensorSpec(shape=(None,len(config.columns["window_metadata"])), dtype=tf.float32),  #windox_X (batch, #wind_x)
         "cl_hits" : tf.TensorSpec(shape=(None,None, None, 4), dtype=tf.float32), #hits  (batch, ncls, nhits, 4)
-        "is_seed": tf.TensorSpec(shape=(None,None), dtype=tf.int64),  # is seed (batch, ncls,)
+        "is_seed": tf.TensorSpec(shape=(None,None), dtype=tf.float32),  # is seed (batch, ncls,)
         "in_scluster": tf.TensorSpec(shape=(None,None), dtype=tf.int64),  # in_supercluster (batch, ncls,)
         "cl_Y": tf.TensorSpec(shape=(None,None, len(config.columns["cl_labels"])), dtype=tf.bool),  #cl_y (batch, ncls, #cl_labels)
         "flavour" : tf.TensorSpec(shape=(None), dtype=tf.float32),  #windox_X (batch, #wind_x)
@@ -141,7 +141,7 @@ def get_tensors_spec(config):
             "wind_X_norm": tf.float32,
             "wind_meta": tf.float32,
             "cl_hits": tf.float32,
-            "is_seed": tf.int64,
+            "is_seed": tf.float32,
             "in_scluster": tf.int64,
             "cl_y": tf.bool,
             "flavour": tf.float32,
